test: move block models to utils
