+++ __tests__/utils/fixtures/unitnet/block-model.ts
@@ -1,0 +1,6 @@
+import { configManager, models } from "@arkecosystem/crypto";
+import genesisBlockJson from "../../config/unitnet/genesisBlock.json";
+
+configManager.setFromPreset("unitnet");
+
+export const genesisBlock = new models.Block(genesisBlockJson);
